+++ build.clj
@@ -5,7 +5,7 @@
             [clojure.spec.alpha :as s]))
 
 (def version (or (System/getenv "VERSION")
-                 "3.15.8"))
+                 "3.16.0"))
 
 (def target-dir "target")
 (def class-dir (str target-dir "/classes"))