+++ CHANGELOG.md
@@ -9,6 +9,8 @@
     
 ### Changes
 
+    - UI refactor for toolbars
+    - UI refactor for printers and multi-thread timeline
     - Set a UI limit of 200 on the Exceptions menu entries to improve UI responsiveness under deep recursions exceptions.
     
 ### Bugs fixed